fold WritableFormatter back in to MsgPackWriter
